+++ lib/src/components/marker/marker.dart
@@ -1,5 +1,4 @@
 import 'dart:ui';
-import '../../values/value.dart';
 import '../component.dart';
 import 'marker_render.dart';
 import 'marker_theme.dart';
@@ -8,30 +7,14 @@ import 'marker_theme.dart';
 class GMarker extends GComponent {
   static const int kDefaultLayer = 1000;
 
-  /// The layer of the marker. highest layer will be on the top.
-  final GValue<int> _layer;
-  int get layer => _layer.value;
-  set layer(int value) => _layer.value = value;
-
-  /// [GHitTestMode] of the marker.
-  final GValue<GHitTestMode> _hitTestMode;
-  GHitTestMode get hitTestMode => _hitTestMode.value;
-  set hitTestMode(GHitTestMode value) => _hitTestMode.value = value;
-
-  /// Whether the marker is highlighted (or selected).
-  final GValue<bool> _highlight = GValue<bool>(false);
-  bool get highlight => _highlight.value;
-  set highlight(bool value) => _highlight.value = value;
-
   GMarker({
     super.id,
     super.visible = true,
     super.theme,
     super.render,
-    int layer = kDefaultLayer,
-    GHitTestMode hitTestMode = GHitTestMode.border,
-  }) : _layer = GValue<int>(layer),
-       _hitTestMode = GValue<GHitTestMode>(hitTestMode);
+    super.layer,
+    super.hitTestMode,
+  });
 
   @override
   GMarkerRender<GMarker, GMarkerTheme> getRender() {