+++ lib/src/graphs/area/area_theme.dart
@@ -5,18 +5,25 @@ import '../../style/paint_style.dart';
 
 /// Theme for area graph
 class GGraphAreaTheme extends GGraphTheme {
-  final PaintStyle styleBaseLine;
-  final PaintStyle styleValueAboveLine;
-  final PaintStyle styleValueBelowLine;
-  final PaintStyle styleAboveArea;
-  final PaintStyle styleBelowArea;
+  /// The style for the area above the base line.
+  ///
+  /// If stroke styles specified the border of the area will be drawn with stroke style.
+  final PaintStyle styleAboveBase;
+
+  /// The style for the area below the base line.
+  ///
+  /// If stroke styles specified the border of the area will be drawn with stroke style.
+  final PaintStyle styleBelowBase;
+
+  /// The style for the base line.
+  ///
+  /// if null, base lines will be drawn with stroke styles in [styleAboveBase] / [styleBelowBase]
+  final PaintStyle? styleBaseLine;
 
   const GGraphAreaTheme({
-    required this.styleBaseLine,
-    required this.styleValueAboveLine,
-    required this.styleValueBelowLine,
-    required this.styleAboveArea,
-    required this.styleBelowArea,
+    required this.styleAboveBase,
+    required this.styleBelowBase,
+    this.styleBaseLine,
     super.axisMarkerTheme,
     super.overlayMarkerTheme,
     super.highlightMarkerTheme,
@@ -24,20 +31,16 @@ class GGraphAreaTheme extends GGraphTheme {
 
   GGraphAreaTheme copyWith({
     PaintStyle? styleBaseLine,
-    PaintStyle? styleValueAboveLine,
-    PaintStyle? styleValueBelowLine,
-    PaintStyle? styleAboveArea,
-    PaintStyle? styleBelowArea,
+    PaintStyle? styleAboveBase,
+    PaintStyle? styleBelowBase,
     GAxisMarkerTheme? axisMarkerTheme,
     GOverlayMarkerTheme? overlayMarkerTheme,
     GGraphHighlightMarkerTheme? highlightMarkerTheme,
   }) {
     return GGraphAreaTheme(
       styleBaseLine: styleBaseLine ?? this.styleBaseLine,
-      styleValueAboveLine: styleValueAboveLine ?? this.styleValueAboveLine,
-      styleValueBelowLine: styleValueBelowLine ?? this.styleValueBelowLine,
-      styleAboveArea: styleAboveArea ?? this.styleAboveArea,
-      styleBelowArea: styleBelowArea ?? this.styleBelowArea,
+      styleAboveBase: styleAboveBase ?? this.styleAboveBase,
+      styleBelowBase: styleBelowBase ?? this.styleBelowBase,
       axisMarkerTheme: axisMarkerTheme ?? this.axisMarkerTheme,
       overlayMarkerTheme: overlayMarkerTheme ?? this.overlayMarkerTheme,
       highlightMarkerTheme: highlightMarkerTheme ?? this.highlightMarkerTheme,