+++ lib/src/graphs/line/line_render.dart
@@ -1,4 +1,6 @@
+import 'dart:typed_data';
 import 'dart:ui';
+import 'dart:ui' as ui;
 
 import 'package:flutter/material.dart';
 
@@ -17,9 +19,8 @@ class GGraphLineRender extends GGraphRender<GGraphLine, GGraphLineTheme> {
     required GValueViewPort valueViewPort,
   }) {
     final dataSource = chart.dataSource;
-    final Path tickLinesPath = Path();
 
-    final List<Vector2> points = <Vector2>[];
+    final List<Vector2> linePoints = <Vector2>[];
     final List<Vector2> highlightMarks = <Vector2>[];
     double highlightInterval = theme.highlightMarkerTheme?.interval ?? 1000.0;
     int highlightIntervalPoints =
@@ -39,35 +40,22 @@ class GGraphLineRender extends GGraphRender<GGraphLine, GGraphLineTheme> {
       }
       double x = pointViewPort.pointToPosition(area, point.toDouble());
       double y = valueViewPort.valueToPosition(area, value);
-      points.add(Vector2(x, y));
+      linePoints.add(Vector2(x, y));
       if (graph.highlight && (point % highlightIntervalPoints == 0)) {
         highlightMarks.add(Vector2(x, y));
       }
     }
-    for (int i = 0; i < points.length - 1; i++) {
-      addLinePath(
-        toPath: tickLinesPath,
-        x1: points[i].x,
-        y1: points[i].y,
-        x2: points[i + 1].x,
-        y2: points[i + 1].y,
-      );
-      if (theme.pointRadius > 0) {
-        Path pointPath = Path();
-        addOvalPath(
-          rect: Rect.fromCircle(
-            center: Offset(points[i].x, points[i].y),
-            radius: theme.pointRadius,
-          ),
-          toPath: pointPath,
-        );
-        drawPath(canvas: canvas, path: pointPath, style: theme.pointStyle);
-      }
-    }
+
+    _drawGraph(
+      canvas: canvas,
+      area: area,
+      theme: theme,
+      linePoints: linePoints,
+    );
+
     if (graph.hitTestMode != GHitTestMode.none) {
-      _hitTestLinePoints.addAll(points);
+      _hitTestLinePoints.addAll(linePoints);
     }
-    drawPath(canvas: canvas, path: tickLinesPath, style: theme.lineStyle);
 
     drawHighlightMarks(
       canvas: canvas,
@@ -75,6 +63,89 @@ class GGraphLineRender extends GGraphRender<GGraphLine, GGraphLineTheme> {
       theme: theme,
       highlightMarks: highlightMarks,
     );
+  }
+
+  // draw graph use raw styles in the theme (optimized batch draw)
+  void _drawGraph({
+    required Canvas canvas,
+    required Rect area,
+    required GGraphLineTheme theme,
+    required List<Vector2> linePoints,
+  }) {
+    Paint? linePaint = theme.lineStyle.getStrokePaint(
+      gradientBounds: theme.lineStyle.gradientBounds ?? area,
+    );
+    Paint? pointStrokePaint = theme.pointStyle.getStrokePaint(
+      gradientBounds: theme.pointStyle.gradientBounds ?? area,
+    );
+    pointStrokePaint?.shader = ui.Gradient.linear(
+      area.topCenter,
+      //the line interval is 20
+      area.bottomCenter,
+      [
+        //give transparent to start color, give actual line color to the end color
+        Colors.transparent,
+        Colors.red,
+      ],
+      //The interval [0.0,0.9] is transparent, and the interval [0.9,1.0] is red. So it looks like drawing red lines
+      [0.90, 1.0],
+      //just repeat the gradient
+      TileMode.repeated,
+      null,
+    );
+    Paint? pointFillPaint = theme.pointStyle.getFillPaint(
+      gradientBounds: theme.pointStyle.gradientBounds ?? area,
+    );
+    if (pointFillPaint != null) {
+      pointFillPaint.strokeWidth = theme.pointRadius * 2;
+      pointFillPaint.style = PaintingStyle.fill;
+      pointFillPaint.strokeCap = pointStrokePaint?.strokeCap ?? StrokeCap.round;
+    }
+    List<double> linesToDraw = [];
+    List<double> pointsToDraw = [];
+    for (int i = 0; i < linePoints.length; i++) {
+      if (linePaint != null) {
+        linesToDraw.add(linePoints[i].x);
+        linesToDraw.add(linePoints[i].y);
+      }
+      if (theme.pointRadius > 0 &&
+          theme.pointStyle.isNotEmpty &&
+          i < linePoints.length - 1) {
+        pointsToDraw.addAll([
+          linePoints[i].x,
+          linePoints[i].y,
+          linePoints[i + 1].x,
+          linePoints[i + 1].y,
+        ]);
+      }
+    }
+    // draw line segments
+    if (linePaint != null) {
+      canvas.drawRawPoints(
+        PointMode.polygon,
+        Float32List.fromList(linesToDraw),
+        linePaint,
+      );
+    }
+    // draw points fill
+    if (pointFillPaint != null) {
+      canvas.drawRawPoints(
+        PointMode.points,
+        Float32List.fromList(pointsToDraw),
+        pointFillPaint,
+      );
+    }
+
+    // draw points stroke
+    if (pointStrokePaint != null) {
+      for (int i = 0; i < linePoints.length; i++) {
+        canvas.drawCircle(
+          Offset(linePoints[i].x, linePoints[i].y),
+          theme.pointRadius,
+          pointStrokePaint,
+        );
+      }
+    }
   }
 
   final List<Vector2> _hitTestLinePoints = [];